smv vv: make some corrections to qfds.sh

+++ Utilities/Scripts/qfds.sh
@@ -205,13 +205,19 @@ if [ $STOPFDS ]; then
  touch $stopfile
  exit
 fi
+if [ "$STOPFDSMAXITER" != "" ]; then
+  echo "creating delayed stop file: $infile"
+  echo $STOPFDSMAXITER > $stopfile
+fi
 if [ "$stopjob" == "1" ]; then
  echo "stopping case: $in"
  touch $stopfile
  exit
 fi
-if [ -e $stopfile ]; then
- rm $stopfile
+if [ "$STOPFDSMAXITER" == "" ]; then
+  if [ -e $stopfile ]; then
+    rm $stopfile
+  fi
 fi
 
 QSUB="qsub -q $queue"
